refactor(dht): copy source codes of parsers to new file
